requirements.txt
requests==2.31.0
geopandas==0.14.1
psycopg2-binary==2.9.9
pyproj==3.6.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
shapely==2.0.2
geojson==3.1.0
jinja2==3.1.2
aiofiles==23.2.1
geoalchemy2==0.14.2
